checking if fixing method for scale method will fix error i encountered when running it

+++ AaliyahAllie-ST10212542-PROG6221-PART2-POE/Recipe.cs
@@ -48,14 +48,14 @@ namespace AaliyahAllie_ST10212542_PROG6221_PART2_POE
 
             Console.WriteLine("**********************************************************************************");
         }
-        
+
         public void ScaleRecipe(double factor)
         {
-            //Loop through each ingredient
+            // Loop through each ingredient
             foreach (var ingredient in Ingredients)
             {
-                //Scale the original quantity of each ingredient by the given factor
-                ingredient.Quantity = (double)(ingredient.OriginalQuantity * factor);
+                // Scale the original quantity of each ingredient by the given factor
+                ingredient.Quantity = (int)(ingredient.OriginalQuantity * factor);
             }
         }
